Use CColRef Id for hash/equal comparisons in Orca
Previously, we would check that each member in a colref is identical
when comparing two colrefs. However, with the introduction of
EUsedStatus, we can end up with two identical colrefs but one labled as
unknown and another as used. This can cause unexpected behavior when
using hashmaps/sets involving colrefs. Instead, use the Id, which is
unique to all colrefs.

Backport of 65276acb0fd8f6f860d017c1fe5bf6aca7220731